Move functions out of image_utils
