+++ README.md
@@ -91,7 +91,7 @@ A short example injecting a simulated object into a stack of images, and then re
 
 ```python
 
-import kbmod.search as kb
+from kbmod.search import StackSearch, Trajectory
 from kbmod.trajectory_generator import KBMODV1Search
 from kbmod.fake_data.fake_data_creator import *
 
@@ -101,7 +101,7 @@ ds = FakeDataSet(512, 512, fake_times)
 
 # Insert an artificial object with starting position x=2, y=0,
 # velocity vx=10.7, vy=15.3, and flux = 275.0.
-trj = kb.Trajectory(x=2, y=0, vx=10.7, vy=15.3, flux=275.0)
+trj = Trajectory(x=2, y=0, vx=10.7, vy=15.3, flux=275.0)
 ds.insert_object(trj)
 
 # Generate a set of trajectories to test from each pixel.
@@ -117,7 +117,7 @@ candidates = [trj for trj in gen]
 
 # Do the actual search (on CPU).  This requires passing in the science
 # images, the variance images, the PSF information, and the times.
-search = kb.StackSearch(
+search = StackSearch(
     ds.stack_py.sci,
     ds.stack_py.var,
     ds.stack_py.psfs,
