+++ src/kbmod/fake_data/fake_data_creator.py
@@ -11,8 +11,9 @@ import numpy as np
 import warnings
 
 from kbmod.configuration import SearchConfiguration
-
-from kbmod.search import *
+from kbmod.core.image_stack_py import ImageStackPy
+from kbmod.core.psf import PSF
+from kbmod.search import Trajectory
 from kbmod.work_unit import WorkUnit
