+++ tests/test_readme_example.py
@@ -1,6 +1,6 @@
 import unittest
 
-import kbmod.search as kb
+from kbmod.search import HAS_GPU, StackSearch, Trajectory
 from kbmod.trajectory_generator import KBMODV1Search
 from kbmod.fake_data.fake_data_creator import *
 
@@ -13,7 +13,7 @@ class test_readme_example(unittest.TestCase):
 
         # Insert an artificial object with starting position x=2, y=0,
         # velocity vx=10.7, vy=15.3, and flux = 275.0.
-        trj = kb.Trajectory(x=2, y=0, vx=10.7, vy=15.3, flux=275.0)
+        trj = Trajectory(x=2, y=0, vx=10.7, vy=15.3, flux=275.0)
         ds.insert_object(trj)
 
         # Generate a set of trajectories to test from each pixel.
@@ -29,7 +29,7 @@ class test_readme_example(unittest.TestCase):
 
         # Do the actual search (on CPU).  This requires passing in the science
         # images, the variance images, the PSF information, and the times.
-        search = kb.StackSearch(
+        search = StackSearch(
             ds.stack_py.sci,
             ds.stack_py.var,
             ds.stack_py.psfs,
